Move test out of autogenerated pb.rs file
